requirements.txt
pipecat-ai[google,silero,deepgram,cartesia,openai]
playwright
python-dotenv
fastapi
uvicorn
chromadb
